+++ src/main/java/com/digimenu/main/domain/projection/PastOrdersProjection.java
@@ -1,10 +1,11 @@
 package com.digimenu.main.domain.projection;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public interface PastOrdersProjection {
     Integer getCount();
-    Double getTotal();
+    BigDecimal getTotal();
     String getName();
     Long getRestaurantId();
     String getRestaurantName();
